+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.2] - 2025-07-31
+
+### Fixed
+
+* fix: Add post deletion handling to update sitemap by @GaryJones in <https://github.com/Automattic/msm-sitemap/pull/211>
+
+### Maintenance
+
+* style: fix more code style violations by @GaryJones in <https://github.com/Automattic/msm-sitemap/pull/212>
+
 ## [1.5.1] - 2025-07-30
 
 ### Added
@@ -182,6 +192,7 @@ Initial release.
 
 * Enable stable Composer installations from origin.
 
+[1.5.2]: https://github.com/automattic/msm-sitemap/compare/1.5.1...1.5.2
 [1.5.1]: https://github.com/automattic/msm-sitemap/compare/1.5.0...1.5.1
 [1.5.0]: https://github.com/automattic/msm-sitemap/compare/1.4.2...1.5.0
 [1.4.2]: https://github.com/automattic/msm-sitemap/compare/1.4.1...1.4.2